refactor directory action code
