+++ client/actions/actionType.js
@@ -5,8 +5,6 @@ export const POST_BUSINESS = 'POST_BUSINESS';
 export const GET_ALL_BUSINESS = 'GET_ALL_BUSINESS';
 export const EDIT_SUCCESSFUL = 'EDIT_SUCCESSFUL';
 export const EDIT_FAILED = 'EDIT_FAILED';
-// export const BUSINESS_DELETE_FAILED = 'BUSINESS_DELETE_FAILED';
-// export const BUSINESS_DELETE_SUCCESSFUL = 'BUSINESS_DELETE_SUCCESSFUL';
 export const DELETE_BUSINESS = 'DELETE_BUSINESS';
 export const GET_SINGLE_BUSINESS = 'GET_SINGLE_BUSINESS';
 export const SAVE_IMAGE_SUCCESSFUL = 'SAVE_IMAGE_SUCCESSFUL';
