+++ client/actions/addBusinessAction.js
@@ -1,0 +1,6 @@
+import axios from 'axios';
+
+export const addBusiness = business => dispatch => {
+    return axios.post('/api/v1/businesses', business)
+        .then(res => res.data.business);
+};
