+++ client/actions/businessAction.js
@@ -1,20 +1,14 @@
 import axios from 'axios';
 import {
   GET_ALL_BUSINESS,
-  SAVE_IMAGE_SUCCESSFUL,
-  SAVE_IMAGE_FAILED,
   GET_SINGLE_BUSINESS,
-  EDIT_FAILED,
-  EDIT_SUCCESSFUL,
-  // BUSINESS_DELETE_FAILED,
-  // BUSINESS_DELETE_SUCCESSFUL
   DELETE_BUSINESS
 } from './actionType';
 
 export function allBusiness(business) {
   return {
-    type: GET_ALL_BUSINESS,
-    allBusinesses: business
+      type: GET_ALL_BUSINESS,
+      allBusinesses: business
   };
 }
 
@@ -25,43 +19,12 @@ export function oneBusiness(business) {
   };
 }
 
-export function editSuccessful(business) {
-  return {
-    type: EDIT_SUCCESSFUL,
-    business
-  };
-}
-
-export function editFailed(error) {
-  return {
-    type: EDIT_FAILED,
-    error
-  };
-}
-
-// export function deleteSuccessful(message) {
-//   return {
-//     type: BUSINESS_DELETE_SUCCESSFUL,
-//     message
-//   };
-// }
-
-// export function deleteFailed(err) {
-//   return {
-//     type: BUSINESS_DELETE_FAILED,
-//     err
-//   };
-// }
 
 const businessToBeDeleted = businessId => ({
   type: DELETE_BUSINESS,
   businessId
  });
 
-export const addBusiness = business => dispatch => {
-  return axios.post('/api/v1/businesses', business)
-    .then(res => res.data.business);
-};
 
 export const getAllBusiness = () => dispatch =>
   axios.get('api/v1/businesses')
@@ -75,56 +38,10 @@ export const getOneBusiness = id => dispatch =>
       dispatch(oneBusiness(res.data));
     });
 
-export const editBusiness = (id, business) => dispatch =>
-  axios.put('/api/v1/businesses/' + id, business)
-    .then(() => {
-      dispatch(editSuccessful(' Business Sucessfully Updated'));
-    })
-    .catch(() => {
-      dispatch(editFailed('Business failed to update'));
-    });
 
-// export const deleteBusiness = id => dispatch =>
-//   axios.delete('/api/v1/businesses/' + id)
-//     .then((res) => {
-//       dispatch(deleteSuccessful(res.data.message));
-//     })
-//     .catch((err) => {
-//       dispatch(deleteFailed(err.res.data.message));
-//     });
 
 export const deleteBusiness = id => dispatch => axios.delete('/api/v1/businesses/' + id)
  .then(() => {
    dispatch(businessToBeDeleted(id));
  });
 
-export function saveImageSuccessful(image) {
-  return {
-    type: SAVE_IMAGE_SUCCESSFUL,
-    image
-  };
-}
-
-export function saveImageFailed(error) {
-  return {
-    type: SAVE_IMAGE_FAILED,
-    error
-  };
-}
-
-export function saveImageCloudinary(image) {
-  const data = new FormData();
-  data.append('file', image);
-  data.append('upload_preset', 'yp3s3k5n');
-  delete axios.defaults.headers.common.Authorization;
-  return dispatch => axios.post('https://api.cloudinary.com/v1_1/dtlziutcq/image/upload', data)
-    .then(({ data }) => {
-      const token = localStorage.getItem('makeToken');
-      //console.log(token);
-      axios.defaults.headers.common.Authorization = token;
-      //console.log(data.secure_url);
-      dispatch(saveImageSuccessful(data.secure_url));
-    }).catch(() => {
-      dispatch(saveImageFailed('Sorry, your image failed to upload'));
-    });
-}
